requirements.txt
# requirements.txt
# =================
# CityCycle London — Python dependencies
#
# Install: pip install -r requirements.txt
#
# Note: google-cloud-bigquery and related packages are optional for
# local mock-data development. They are only needed when USE_MOCK_DATA=false.

# ── Data engineering ──────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.25.0
pyarrow>=12.0.0              # Required by pandas BQ integration

# ── BigQuery (optional for dev — required for prod) ───────────────
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-storage>=2.22.0   # faster BQ downloads
sqlalchemy>=2.0.0
sqlalchemy-bigquery>=1.8.0              # SQLAlchemy BQ adapter (Jupyter + GE)
db-dtypes>=1.1.0                        # BQ type handling

# ── dbt ───────────────────────────────────────────────────────────
dbt-core>=1.7.0
dbt-bigquery>=1.7.0

# ── Meltano ───────────────────────────────────────────────────────
meltano>=3.3.0

# ── Great Expectations ────────────────────────────────────────────
great-expectations>=0.18.0

# ── Dagster ───────────────────────────────────────────────────────
dagster>=1.6.0
dagster-webserver>=1.6.0

# ── ML ────────────────────────────────────────────────────────────
scikit-learn>=1.3.0
joblib>=1.3.0
xgboost>=2.0.0               # alternative to RandomForest

# ── Dashboard ─────────────────────────────────────────────────────
streamlit>=1.32.0
pydeck>=0.8.0                # geospatial map
plotly>=5.18.0               # optional: richer charts

# ── Utilities ─────────────────────────────────────────────────────
python-dotenv>=1.0.0         # load .env file
faker>=20.0.0                # used in mock data generator
tqdm>=4.66.0                 # progress bars
